+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Version 4.6.11-beta (not yet released)
 
+**Other improvements and fixes**
+- Fix some weather texts not being updated after a forced refresh after a language change
+
 **Translations**
 - Translations updated